requirements.txt
asgineer==0.9.4
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
h11==0.14.0
h5py==3.13.0
harp-protocol==0.3.0
harp-serial==0.3.0
harp.laserdrivercontroller==0.1.0a2
iptools==0.7.0
itemdb==1.3.0
Jinja2==3.1.6
Markdown==3.7
MarkupSafe==3.0.2
numpy==1.26.4
opencv-python==4.11.0.86
pscript==0.8.0
PyJWT==2.10.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.9.3.3
pyqtgraph==0.13.7
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
spinnaker-python @ file:///C:/Users/bioph/Projects/pyProbe/spinnaker_python-3.2.0.62-cp310-cp310-win_amd64.whl#sha256=6a998a1822cb9f9b131f16b8e9b63dbc9394286af876555351de8e4bdc3417a6
timetagger==24.12.2
uvicorn==0.34.0
